Applied patch from jira issue - OFBIZ-5766 - Error when adding/modifying a task to a Production Run - field type error. Thanks Guillaume Sigoigne for reporting the issue, Thanks Divesh for providing the patch for this issue.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1639880 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/manufacturing/src/org/ofbiz/manufacturing/jobshopmgt/ProductionRun.java
@@ -423,7 +423,7 @@ public class ProductionRun {
                     Map<String, Object> serviceContext = UtilMisc.<String, Object>toMap("arguments", estimateCalcServiceMap);
                     // serviceContext.put("userLogin", userLogin);
                     Map<String, Object> resultService = dispatcher.runSync(serviceName, serviceContext);
-                    totalTaskTime = ((Double)resultService.get("totalTime")).doubleValue();
+                    totalTaskTime = ((BigDecimal)resultService.get("totalTime")).doubleValue();
                 }
             } catch (Exception exc) {
                 Debug.logError(exc, "Problem calling the customMethod service " + serviceName);
